refactor(test) update htjml2js settings
