+++ .babelrc
@@ -1,6 +1,3 @@
 {
-  "presets": ["es2015"],
-  "plugins": [
-    "transform-flow-strip-types"
-  ]
+  "presets": ["env", "flow"]
 }
