+++ package.json
@@ -14,16 +14,18 @@
     "@uncovertruth/stylelint-config": "^0.2.0",
     "babel-core": "^6.24.1",
     "babel-loader": "^7.0.0",
-    "babel-plugin-transform-flow-strip-types": "^6.22.0",
-    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-flow": "^6.23.0",
     "babel-register": "^6.24.1",
+    "cross-env": "^4.0.0",
+    "css-loader": "^0.28.1",
     "del": "^2.2.2",
     "eslint": "^3.19.0",
     "flow-bin": "^0.45.0",
     "gulp": "^3.9.1",
     "gulp-babel": "^6.1.2",
     "gulp-cache": "^0.4.6",
-    "gulp-chrome-manifest": "0.0.13",
+    "gulp-chrome-manifest": "^0.0.13",
     "gulp-cssnano": "^2.1.2",
     "gulp-debug": "^3.1.0",
     "gulp-htmlmin": "^3.0.0",
@@ -47,7 +49,7 @@
     "npm-run-all": "^4.0.2",
     "power-assert": "^1.4.2",
     "pug-lint": "^2.4.0",
-    "pug-lint-config-uncovertruth": "0.0.2",
+    "pug-lint-config-uncovertruth": "^0.0.2",
     "rimraf": "^2.6.1",
     "run-sequence": "^1.2.2",
     "stylelint": "^7.10.1",
@@ -68,7 +70,7 @@
   "scripts": {
     "build": "npm-run-all -p build:*",
     "build:gulp": "gulp build",
-    "build:webpack": "webpack",
+    "build:webpack": "cross-env NODE_ENV=production webpack",
     "clean": "find app -name '*.html' -delete && rimraf dist app/scripts package",
     "lint": "npm-run-all -p lint:*",
     "lint:css": "stylelint app/styles/*.css",
@@ -79,7 +81,7 @@
     "release": "gulp package",
     "tag": "uglifyjs app/scripts.babel/tag.js -m",
     "test": "npm-run-all test:*",
-    "test:karma": "karma start test/karma.conf.js --single-run",
+    "test:karma": "karma start --single-run",
     "test:lint": "npm run lint",
     "watch": "gulp watch"
   }
