+++ test/index.test.js
@@ -1,16 +1,17 @@
 /* @flow */
-import assert from 'assert';
-import Render from '../app/render';
-import { describe, it, before } from 'mocha';
+import { describe, it, before } from 'mocha'
+import assert from 'assert'
+
+import Render from '../app/render'
 
 describe('Vue unit test', () => {
   before(() => {
-    document.body.innerHTML = window.__html__['fixtures/test.html'];
-    Render();
-  });
+    document.body.innerHTML = window.__html__['test/fixtures/test.html']
+    Render()
+  })
 
   it('should render dom', () => {
-    const dom = document.getElementsByTagName('td')[0];
-    assert.equal(dom.innerHTML, 'key');
-  });
-});
+    const dom = document.getElementsByTagName('td')[0]
+    assert.equal(dom.innerHTML, 'key')
+  })
+})
