--- test/karma.conf.js
@@ -1,2 +1,0 @@
-require('babel-register');
-module.exports = require('./karma.conf.babel').default;
