--- test/mocha.opts
@@ -1,2 +1,0 @@
---compilers js:babel-register
---ui bdd
